refactor: genericize the WHIR-R1CS scheme builder in the spine
